linter

+++ examples/ex09/Punct_stream/test_ps.cpp
@@ -22,7 +22,7 @@ int main()
   }
 
   std::sort(vs.begin(), vs.end());
-  for (int i = 0; i < vs.size(); ++i) {
+  for (size_t i = 0; i < vs.size(); ++i) {
     if (i == 0 || vs[i] != vs[i - 1])
       cout << vs[i] << '\n';
   }
